# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for FanFic AI
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories/{id}/tags:
    get:
      operationId: getStoryTags
      tags: [stories]
      summary: Get tags attached to a story
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Story tags
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tag" }
    put:
      operationId: setStoryTags
      tags: [stories]
      summary: Replace the tag set of a story (author only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateStoryTagsBody" }
      responses:
        "200":
          description: Updated tags
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tag" }
...
  /series:
    get:
      operationId: listSeries
      tags: [stories]
      summary: List series, optionally filtered by author
      parameters:
        - name: authorName
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Series
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Series" }
    post:
      operationId: createSeries
      tags: [stories]
      summary: Create a series
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSeriesBody" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Series" } } } }
  /series/{id}:
    get:
      operationId: getSeries
      tags: [stories]
      summary: Get a series with its stories
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Series, content: { application/json: { schema: { $ref: "#/components/schemas/SeriesWithStories" } } } }
    patch:
      operationId: updateSeries
      tags: [stories]
      summary: Update title/summary
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSeriesBody" }
      responses:
        "200": { description: Series, content: { application/json: { schema: { $ref: "#/components/schemas/Series" } } } }
    delete:
      operationId: deleteSeries
      tags: [stories]
      summary: Delete a series
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: requesterAuthorName
          in: query
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
  /series/{id}/stories:
    post:
      operationId: addStoryToSeries
      tags: [stories]
      summary: Add a story to a series (author only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddStoryToSeriesBody" }
      responses:
        "200": { description: Series, content: { application/json: { schema: { $ref: "#/components/schemas/SeriesWithStories" } } } }
  /series/{id}/stories/{storyId}:
    delete:
      operationId: removeStoryFromSeries
      tags: [stories]
      summary: Remove a story from a series
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: storyId
          in: path
          required: true
          schema: { type: integer }
        - name: requesterAuthorName
          in: query
          required: true
          schema: { type: string }
      responses:
        "204": { description: Removed }
  /authors/{name}/notification-prefs:
    get:
      operationId: getNotificationPrefs
      tags: [notifications]
      summary: Get an author's notification preferences
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Prefs, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationPrefs" } } } }
    put:
      operationId: updateNotificationPrefs
      tags: [notifications]
      summary: Update an author's notification preferences
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateNotificationPrefsBody" }
      responses:
        "200": { description: Prefs, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationPrefs" } } } }
components:
  schemas:
...
    Story:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        genre:
          type: string
        artStyle:
          type: string
        lengthSetting:
          type: string
          enum: [short, medium, long]
        seedPrompt:
          type: ["string", "null"]
        fullText:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        characters:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published]
        authorName:
          type: string
        coAuthors:
          type: array
          items:
            type: string
          default: []
        coverImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
      required:
        - id
        - title
        - genre
        - artStyle
        - lengthSetting
        - status
        - coAuthors
        - authorName
        - createdAt
        - updatedAt
        - likeCount
        - commentCount
...
    Tag:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        label: { type: string }
        storyCount: { type: integer }
      required: [id, slug, label, storyCount]
    UpdateStoryTagsBody:
      type: object
      properties:
        slugs:
          type: array
          items: { type: string, minLength: 1, maxLength: 30 }
        requesterAuthorName: { type: string, minLength: 1 }
      required: [slugs, requesterAuthorName]
...
    Series:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        summary: { type: ["string", "null"] }
        authorName: { type: string }
        storyCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, title, authorName, storyCount, createdAt, updatedAt]
    SeriesWithStories:
      allOf:
        - $ref: "#/components/schemas/Series"
        - type: object
          properties:
            stories:
              type: array
              items:
                allOf:
                  - $ref: "#/components/schemas/Story"
                  - type: object
                    properties:
                      position: { type: integer }
                    required: [position]
          required: [stories]
    CreateSeriesBody:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        summary: { type: ["string", "null"] }
        authorName: { type: string, minLength: 1 }
      required: [title, authorName]
    UpdateSeriesBody:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        summary: { type: ["string", "null"] }
        requesterAuthorName: { type: string, minLength: 1 }
      required: [requesterAuthorName]
    AddStoryToSeriesBody:
      type: object
      properties:
        storyId: { type: integer }
        position: { type: integer, default: 0 }
        requesterAuthorName: { type: string, minLength: 1 }
      required: [storyId, requesterAuthorName]
    NotificationPrefs:
      type: object
      properties:
        authorName: { type: string }
        comment: { type: boolean }
        follow: { type: boolean }
        like: { type: boolean }
        repost: { type: boolean }
        coAuthorChapter: { type: boolean }
      required: [authorName, comment, follow, like, repost, coAuthorChapter]
    UpdateNotificationPrefsBody:
      type: object
      properties:
        comment: { type: boolean }
        follow: { type: boolean }
        like: { type: boolean }
        repost: { type: boolean }
        coAuthorChapter: { type: boolean }
      additionalProperties: false
...